Adds back Base64 interface
